% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/4_NetworkModels.R
\name{calculateG}
\alias{calculateG}
\title{Computes matrix G}
\usage{
calculateG(ps)
}
\arguments{
\item{ps}{partial solution of closed network}
}
\value{
matrix G
}
\description{
Computes matrix G
}
\keyword{internal}

