% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cdfcompggplot2}
\alias{cdfcompggplot2}
\title{Cumulative Density plot using the package ggplot2}
\usage{
cdfcompggplot2(lfitdata)
}
\arguments{
\item{lfitdata}{a list of fitted data}
}
\description{
Cumulative Density plot using the package ggplot2
}
\seealso{
Other DistributionAnalysis: \code{\link{denscompggplot2}};
  \code{\link{fitData}}; \code{\link{goodnessFit}};
  \code{\link{qqcompggplot2}}; \code{\link{summaryFit}}
}

