% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/6_DistributionAnalysis.R
\name{chisq.test.cont}
\alias{chisq.test.cont}
\title{Computes the chi-square fitness test for a continuous distribution, splited in equal probabilities intervals.}
\usage{
chisq.test.cont(x, distribution = "norm", nclasses = min(100, max(3,
  floor(length(x)/5))), output = FALSE, nestpar = 0, ...)
}
\arguments{
\item{x}{list of data}

\item{distribution}{string that indicates the distribution to fit}

\item{nclasses}{number of intervals}

\item{output}{if is TRUE, prints the chi-square table}

\item{nestpar}{numbr of estimated parameters}

\item{...}{distribution parameters}
}
\value{
the computed statistic and the p-value
}
\description{
Computes the chi-square fitness test for a continuous distribution, splited in equal probabilities intervals.
}
\keyword{internal}

