% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NetworkModels.R
\name{CN_example}
\alias{CN_example}
\title{Data for a Closed Network Example}
\usage{
CN_example()
}
\value{
The solution of the example
}
\description{
Data for a Closed Network Example
}
\keyword{internal}

