% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DistributionAnalysis.R
\name{denscompggplot2}
\alias{denscompggplot2}
\title{Density histogram Plot using the package ggplot2}
\usage{
denscompggplot2(lfitdata)
}
\arguments{
\item{lfitdata}{a list of data fitted}
}
\description{
Density histogram Plot using the package ggplot2
}
\seealso{
Other DistributionAnalysis: \code{\link{cdfcompggplot2}};
  \code{\link{fitData}}; \code{\link{goodnessFit}};
  \code{\link{qqcompggplot2}}; \code{\link{summaryFit}}
}

