% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/4_NetworkModels.R
\name{f_close}
\alias{f_close}
\title{Computes \ifelse{latex}{\eqn{f_{i}(n)}}{\out{<i>f<sub>i</sub>(n)</i>}}}
\usage{
f_close(ps, i, n)
}
\arguments{
\item{ps}{partial solution of closed network}

\item{i}{node}

\item{n}{clients}
}
\value{
\ifelse{latex}{\eqn{f_{i}(n)}}{\out{<i>f<sub<i</sub>(n)</i>}}
}
\description{
Computes \ifelse{latex}{\eqn{f_{i}(n)}}{\out{<i>f<sub>i</sub>(n)</i>}}
}
\keyword{internal}

