% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/6_DistributionAnalysis.R
\name{fitData}
\alias{fitData}
\title{Computes the estimated parameters of the distributions for the input data}
\usage{
fitData(data, ldistr = c("exp", "norm", "weibull", "unif", "lnorm", "gamma",
  "beta"))
}
\arguments{
\item{data}{data to estimate parameters}

\item{ldistr}{A list of distributions}
}
\value{
A list of estimated parameters for each distribution
}
\description{
Computes the estimated parameters of the distributions for the input data
}
\examples{
mydata <- rnorm(100, 10, 0.5)

fitData(mydata)
}
\seealso{
Other DistributionAnalysis: \code{\link{cdfcompggplot2}};
  \code{\link{denscompggplot2}}; \code{\link{goodnessFit}};
  \code{\link{qqcompggplot2}}; \code{\link{summaryFit}}
}

