% Generated by roxygen2 (4.0.2): do not edit by hand
\name{M_M_S}
\alias{M_M_S}
\title{Obtains the main characteristics of a M/M/s queueing model}
\usage{
M_M_S(lambda = 3, mu = 6, s = 2)
}
\arguments{
\item{lambda}{Mean arrival rate}

\item{mu}{Mean service rate}

\item{s}{Number of servers}
}
\value{
Returns the next information of a M/M/s model:
\item{rho}{Traffic intensity: \eqn{\rho}}
\item{cn}{Coefficients used in the computation of \ifelse{latex}{\eqn{P_n}: \eqn{C_{n}}}{\out{<i>P<sub>n</sub></i>: <i>C<sub>n</sub></i>}}}
\item{p0}{Probability of empty system: \ifelse{latex}{\eqn{P_{0}}}{\out{<i>P<sub>0</sub></i>}}}
\item{l}{Number of customers in the system:  \eqn{L}}
\item{lq}{Number of customers in the queue: \ifelse{latex}{\eqn{L_{q}}}{\out{<i>L<sub>0</sub></i>}}}
\item{w}{Waiting time in the system: \eqn{W}}
\item{wq}{Waiting time in the queue: \ifelse{latex}{\eqn{W_{q}}}{\out{<i>W<sub>0</sub></i>}}}
\item{eff}{System efficiency: \ifelse{latex}{\eqn{Eff = W/(W - W_q)}}{\out{<i>Eff = W/(W - W<sub>q</sub></i>)}}}
}
\description{
Obtains the main characteristics of a M/M/s queueing model
}
\examples{
M_M_S(60, 20, 4)
}
\seealso{
Other AnaliticalModels: \code{\link{ClosedJacksonNetwork}};
  \code{\link{M_M_1_INF_H}}; \code{\link{M_M_1_K}};
  \code{\link{M_M_1}}; \code{\link{M_M_INF}};
  \code{\link{M_M_S_INF_H_Y}}; \code{\link{M_M_S_INF_H}};
  \code{\link{M_M_S_K}}; \code{\link{OpenJacksonNetwork}}
}

