% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ModelsHeader.R
\name{MarkovianModel}
\alias{MarkovianModel}
\title{Defines a queueing model}
\usage{
MarkovianModel(arrivalDistribution = Exp(1), serviceDistribution = Exp(1))
}
\arguments{
\item{arrivalDistribution}{Arrival distribution (object of S4-class \code{distr}
defined in \pkg{distr} package)}

\item{serviceDistribution}{Service distribution (object of S4-class \code{distr}
defined in \pkg{distr} package)}
}
\value{
An object of class \code{MarkovianModel}, a list with the following components:
\item{arrivalDistribution}{Arrival distribution (object of S4-class \code{distr}
defined in \pkg{distr} package)}
\item{serviceDistribution}{Service distribution (object of S4-class \code{distr}
defined in \pkg{distr} package)}
}
\description{
Constructor for \code{MarkovianModel} class.
}

