% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1_ModelsHeader.R
\name{no_distr}
\alias{no_distr}
\title{Defines an empty object representing the inexistence of a distribution.}
\usage{
no_distr()
}
\description{
Defines an empty object representing the inexistence of a distribution.
}

