% Generated by roxygen2 (4.0.2): do not edit by hand
\name{node.OpenJackson}
\alias{node.OpenJackson}
\title{Gets the model of the selected node}
\usage{
\method{node}{OpenJackson}(net, i)
}
\arguments{
\item{net}{network}

\item{i}{node}
}
\value{
object model
}
\description{
Gets the model of the selected node
}
\keyword{internal}

