% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1_ModelsHeader.R, R/4_NetworkModels.R
\name{P0i}
\alias{P0i}
\alias{P0i.OpenJackson}
\title{Steady-state probability of 0 customers in the system on the node i of an Open Jackson Network.}
\usage{
P0i(net, i)

\method{P0i}{OpenJackson}(net, i)
}
\arguments{
\item{net}{Network}

\item{i}{Node. Index starts in 1.}
}
\value{
\ifelse{latex}{\deqn{P_{0,i}}}{\out{<i>P<sub>0,i</sub></i>}}
}
\description{
Returns the value of the probability of having 0 customers at node i of an Open Jackson Network.
}
\details{
\code{P0i.OpenJackson} implements the method for an Open Jackson Network model
}
\examples{
#Probability of having 0 customers on the node 2
P0i(OpenJacksonNetwork(), 2)
}

