% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/5_SimulateModels.R
\name{ParallelizeSimulations}
\alias{ParallelizeSimulations}
\title{Parallelize an queue model function between multiple processors}
\usage{
ParallelizeSimulations(modelfunction, parameters, nsim, nproc)
}
\arguments{
\item{modelfunction}{Function that implements a simulated queue model}

\item{parameters}{List of parameters of the simulated queue model}

\item{nsim}{Number of simulations}

\item{nproc}{Processors to use in the simulation.}
}
\value{
The main characteristcs of the given model
}
\description{
Parallelize an queue model function between multiple processors
}
\keyword{internal}

