% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1_ModelsHeader.R, R/4_NetworkModels.R
\name{Pi0}
\alias{Pi0}
\alias{Pi0.OpenJackson}
\title{Steady-state probability of i customers in the system on the node 0 of an Open Jackson Network.}
\usage{
Pi0(net, i)

\method{Pi0}{OpenJackson}(net, i)
}
\arguments{
\item{net}{Open Jackson Network}

\item{i}{Customers}
}
\value{
\ifelse{latex}{\deqn{P_{i,0}}}{\out{<i>P<sub>i,0</sub></i>}}
}
\description{
Returns the value of the probability of i customers in node 0 of an Open Jackson Network.
}
\details{
\code{Pi0.OpenJackson} implements the method for an Open Jackson Network model
}
\keyword{internal}

