% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1_ModelsHeader.R
\name{plot.MarkovianModel}
\alias{plot.MarkovianModel}
\title{Shows the main graphics of the parameters of a Markovian Model}
\usage{
\method{plot}{MarkovianModel}(x, t = list(range = seq(x$out$w, x$out$w * 3,
  length.out = 100)), n = c(0:5), only = NULL, graphics = "ggplot2", ...)
}
\arguments{
\item{x}{Markovian Model}

\item{t}{range for drawing the waiting plots}

\item{n}{range for drawing the probabilities plot}

\item{only}{Allow to only show the waiting plots or the probabilites plots.
Must be NULL, "t" or "n"}

\item{graphics}{library used to draw the plots}

\item{...}{Further arguments}
}
\description{
Shows the main graphics of the parameters of a Markovian Model
}
\details{
\code{plot.MarkovianModel} implements the function for an object of class MarkovianModel.
}

