% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NetworkModels.R
\name{pnlast}
\alias{pnlast}
\title{Probability of i clients in the last node of the network}
\usage{
pnlast(ps, i)
}
\arguments{
\item{ps}{partial solution of closed network}

\item{i}{clients}
}
\value{
P(i) in the last node
}
\description{
Probability of i clients in the last node of the network
}
\keyword{internal}

