\name{print.Closed}
\alias{print.Closed}
\title{Print the main characteristics of a Closed object}
\usage{
  \method{print}{Closed} (x, ...)
}
\arguments{
  \item{x}{Closed object}

  \item{...}{Further arguments passed to or from other
  methods.}
}
\description{
  Print the main characteristics of a Closed object
}
\keyword{internal}

