\name{print.MarkovianModel}
\alias{print.MarkovianModel}
\title{Print the main characteristics of a queueing model}
\usage{
  \method{print}{MarkovianModel} (x, ...)
}
\arguments{
  \item{x}{MarkovianModel object}

  \item{...}{Further arguments passed to or from other
  methods.}
}
\description{
  Print the main characteristics of a queueing model
}
\keyword{internal}

