\name{print.Open}
\alias{print.Open}
\title{Print the main characteristics of a Open object}
\usage{
  \method{print}{Open} (x, ...)
}
\arguments{
  \item{x}{Open object}

  \item{...}{Further arguments passed to or from other
  methods.}
}
\description{
  Print the main characteristics of a Open object
}
\keyword{internal}

