% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.OpenJackson}
\alias{print.OpenJackson}
\title{Print the main characteristics of a Open Jackson Network model}
\usage{
\method{print}{OpenJackson}(x, ...)
}
\arguments{
\item{x}{a OpenJackson object}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print the main characteristics of a Open Jackson Network model
}
\keyword{internal}

