% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/4_NetworkModels.R
\name{SN_Example}
\alias{SN_Example}
\title{Data for a Open Network Example}
\usage{
SN_Example()
}
\value{
The solution of the example
}
\description{
Data for a Open Network Example
}
\keyword{internal}

