% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/5_SimulateModels.R
\name{summary.SimulatedModel}
\alias{summary.SimulatedModel}
\title{Shows basic statistical information of the variable}
\usage{
\method{summary}{SimulatedModel}(object, var = "l", ...)
}
\arguments{
\item{object}{SimulatedModel}

\item{var}{Main characteristic of an queue model}

\item{...}{Further argumets}
}
\description{
Shows basic statistical information of the variable
}
\details{
\code{summary.SimulatedModel} implements the function for an object of class SimulatedModel.
}

