% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summaryFit}
\alias{summaryFit}
\title{Shows three plots:
     The histogram and theoretical densities
     The empirical and thoeretical CDF's
     The Q-Q plot}
\usage{
summaryFit(lfitdata, graphics = "ggplot2")
}
\arguments{
\item{lfitdata}{a list of fitted data}

\item{graphics}{Type of graphics: "graphics" use the basic R plot and "ggplot2" the library ggplot2}
}
\description{
Shows three plots:
     The histogram and theoretical densities
     The empirical and thoeretical CDF's
     The Q-Q plot
}
\seealso{
Other DistributionAnalysis: \code{\link{cdfcompggplot2}};
  \code{\link{denscompggplot2}}; \code{\link{fitData}};
  \code{\link{goodnessFit}}; \code{\link{qqcompggplot2}}
}

