% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summarySimple}
\alias{summarySimple}
\alias{summarySimple.SimulatedModel}
\alias{summarySimple.SimulatedNetwork}
\alias{summarySimple.list}
\title{Shows a plot of the evolution of a variable during the simulation}
\usage{
summarySimple(object, minrange, maxrange, var, graphics, depth)

\method{summarySimple}{SimulatedModel}(object, minrange, maxrange, var,
  graphics = "ggplot2", depth = maxrange - minrange)

\method{summarySimple}{SimulatedNetwork}(object, minrange, maxrange, var,
  graphics = "ggplot2", depth = maxrange - minrange)

\method{summarySimple}{list}(object, minrange, maxrange, var,
  graphics = "ggplot2", depth = 100)
}
\arguments{
\item{object}{Simulated Model}

\item{minrange}{First client choosen in the simulation}

\item{maxrange}{Last client choosen in the simulation}

\item{var}{Variable to show graphic (L, Lq, W, Wq)}

\item{graphics}{Type of graphics: "graphics" use the basic R plot and "ggplot2" the library ggplot2}

\item{depth}{Number of point printed in the plot}
}
\description{
Shows a plot of the evolution of a variable during the simulation
}
\details{
\code{summarySimple.SimulatedModel} implements the function for an object of class SimulatedModel.

\code{summarySimple.SimulatedNetwork} implements the function for an object of class SimulatedNetwork.

\code{summarySimple.list} implements the function for an object of class list
}

