% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1_ModelsHeader.R
\name{summaryWtWqt}
\alias{summaryWtWqt}
\title{Plot of the waiting times of a Markovian Model}
\usage{
summaryWtWqt(object, t, graphics = "ggplot2")
}
\arguments{
\item{object}{Markovian Model}

\item{t}{Range of t}

\item{graphics}{Type of graphics: "graphics" uses the basic R plot and "ggplot2" the library ggplot2}
}
\description{
Shows a plot with the W(t) and \ifelse{latex}{\eqn{W_{q}(t)}}{<i>W<sub>q</sub>(t)<i>} values of a Markovian Model
}
\keyword{internal}

