\name{summarywwq}
\alias{summarywwq}
\title{Shows a plot of the evolution of W and Wq during the simulation}
\usage{
  summarywwq(object, minrange, maxrange,
    graphics = "ggplot2")
}
\arguments{
  \item{object}{Simulated Model}

  \item{minrange}{Start client of simulation}

  \item{maxrange}{Last client of simulation}

  \item{graphics}{Type of graphics: "graphics" use the
  basic R plot and "ggplot2" the library ggplot2}
}
\description{
  Shows a plot of the evolution of W and Wq during the
  simulation
}

