\docType{data}
\name{uiList}
\alias{uiList}
\title{List of export models to the UI}
\format{List of 20
 $ :List of 4
  ..- attr(*, "class")= chr "markovian"
 $ :List of 4
  ..- attr(*, "class")= chr "markovian"
 $ :List of 4
  ..- attr(*, "class")= chr "markovian"
 $ :List of 4
  ..- attr(*, "class")= chr "markovian"
 $ :List of 4
  ..- attr(*, "class")= chr "markovian"
 $ :List of 4
  ..- attr(*, "class")= chr "markovian"
 $ :List of 4
  ..- attr(*, "class")= chr "markovian"
 $ :List of 4
  ..- attr(*, "class")= chr "markovian"
 $ :List of 4
  ..- attr(*, "class")= chr [1:2] "network" "markovian"
 $ :List of 4
  ..- attr(*, "class")= chr [1:2] "network" "markovian"
 $ :List of 4
  ..- attr(*, "class")= chr "simulation"
 $ :List of 4
  ..- attr(*, "class")= chr "simulation"
 $ :List of 4
  ..- attr(*, "class")= chr "simulation"
 $ :List of 4
  ..- attr(*, "class")= chr "simulation"
 $ :List of 4
  ..- attr(*, "class")= chr "simulation"
 $ :List of 4
  ..- attr(*, "class")= chr "simulation"
 $ :List of 4
  ..- attr(*, "class")= chr "simulation"
 $ :List of 4
  ..- attr(*, "class")= chr "simulation"
 $ :List of 4
  ..- attr(*, "class")= chr "simulation"
 $ :List of 4
  ..- attr(*, "class")= chr "simulation"}
\usage{
  uiList
}
\description{
  List of export models to the UI
}
\keyword{datasets}
\keyword{internal}

