% Generated by roxygen2 (4.0.2): do not edit by hand
\name{combineSimulations}
\alias{combineSimulations}
\title{Combines a list of independent simulations of a queueing model, calculating the mean and variance of each characteristic of interest}
\usage{
combineSimulations(listsims)
}
\arguments{
\item{listsims}{A list of independent simulations}
}
\value{
an object with the mean and variance of parameters L, Lq, W, Wq, Rho and Eff.
}
\description{
Combines a list of independent simulations of a queueing model, calculating the mean and variance of each characteristic of interest
}
\examples{
combineSimulations(G_G_1(nsim=5))
}

