% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fitData}
\alias{fitData}
\title{Computes the estimated parameters of the giving distributions for the input data}
\usage{
fitData(data, ldistr = c("exp", "norm", "weibull", "unif", "lnorm", "gamma"))
}
\arguments{
\item{data}{data to estimate parameters}

\item{ldistr}{A list of distributions names for estimate its parameters}
}
\value{
A list of estimate parameters for each distribution
}
\description{
Computes the estimated parameters of the giving distributions for the input data
}
\seealso{
Other DistributionAnalysis: \code{\link{cdfcompggplot2}};
  \code{\link{denscompggplot2}}; \code{\link{goodnessFit}};
  \code{\link{qqcompggplot2}}; \code{\link{summaryFit}}
}

