% Generated by roxygen2 (4.0.2): do not edit by hand
\name{FWq.M_M_1_INF_H}
\alias{FWq}
\alias{FWq.M_M_1}
\alias{FWq.M_M_1_INF_H}
\alias{FWq.M_M_1_K}
\alias{FWq.M_M_INF}
\alias{FWq.M_M_S}
\alias{FWq.M_M_S_INF_H}
\alias{FWq.M_M_S_INF_H_Y}
\alias{FWq.M_M_S_K}
\alias{FWq.MarkovianModel}
\title{Distribution function of the waiting time in the queue}
\usage{
\method{FWq}{M_M_1_INF_H}(qm, x)

\method{FWq}{M_M_S_INF_H}(qm, x)

\method{FWq}{M_M_S_INF_H_Y}(qm, x)

\method{FWq}{M_M_INF}(qm, x)

FWq(qm, x)

\method{FWq}{MarkovianModel}(qm, x)

\method{FWq}{M_M_1}(qm, x)

\method{FWq}{M_M_S}(qm, x)

\method{FWq}{M_M_1_K}(qm, x)

\method{FWq}{M_M_S_K}(qm, x)
}
\arguments{
\item{qm}{Queueing model}

\item{x}{Time}
}
\value{
\ifelse{latex}{\deqn{W_{q}(x)}}{\out{<i>W<sub>q</sub>(x)</i>}}
}
\description{
Returns the value of the cumulative distribution function of waiting time in the queue
 \ifelse{latex}{\deqn{W_{q} = P(W_{q} \le x)}}{\out{<br><center><i>W<sub>q</sub> = P(W<sub>q</sub> &le; x)</i></center>}}
}
\details{
\code{FWq.M_M_1_INF_H} implements the method for a M/M/1/\eqn{\infty}/H queueing model

\code{FWq.M_M_S_INF_H} implements the method for a M/M/s/\eqn{\infty}/H queueing model

\code{FWq.M_M_S_INF_H_Y} implements the method for a M/M/s/\eqn{\infty}/H with Y replacements queueing model

\code{FWq.M_M_INF} implements the method for a M/M/\eqn{\infty} queueing model

\code{FWq.MarkovianModel} implements the default method (generates a message)

\code{FWq.M_M_1} implements the method for a M/M/1 queueing model

\code{FWq.M_M_S} implements the method for a M/M/S queueing model

\code{FWq.M_M_1_K} implements the method for a M/M/1/K queueing model

\code{FWq.M_M_S_K} implements the method for a M/M/S/K queueing model
}
\examples{
qm <- M_M_1_INF_H(20, 60, 10)
#Cumulative probabilites of waiting 0, 0.25 and 0.5 unit of times in the queue
FWq(qm, c(0, 0.25, 0.5))
qm <- M_M_S_INF_H(10, 40, 5, 8)
#Cumulative probability of waiting 0, 0.25 and 0.5 units of time in the queue
FWq(qm, c(0, 0.25, 0.5))
qm <- M_M_S_INF_H_Y(3, 6, 2, 5, 8)
#Cumulative probability of waiting 0, 0.25 and 0.5 units of time in the queue
FWq(qm, c(0, 0.25, 0.5))
qm <- M_M_INF(15, 23)
#Cumulative probability of waiting 0, 0.25 and 0.5 units of time in the queue
FWq(qm, c(0, 0.25, 0.5))
qm <- M_M_1(60, 80)
#Cumulative probabilites of waiting 0, 0.25 and 0.5 unit of times in the queue
FWq(qm, c(0, 0.25, 0.5))
qm <- M_M_S(60, 20, 4)
#Cumulative probability of waiting 0, 0.25 and 0.5 units of time in the system
FWq(qm, c(0, 0.25, 0.5))
qm <- M_M_1_K(60, 10, 3)
#Cumulative probability of waiting 0, 0.25 and 0.5 units of time in the queue
FWq(qm, c(0, 0.25, 0.5))
qm <- M_M_S_K(60, 10, 2, 5)
#Cumulative probability of waiting 0, 0.25 and 0.5 units of time in the queue
FWq(qm, c(0, 0.25, 0.5))
}

