% Generated by roxygen2 (4.0.2): do not edit by hand
\name{goodnessFit}
\alias{goodnessFit}
\title{Computes the p-value of the chi-square test and Kolmogorov-Smirnov test and each statistic value}
\usage{
goodnessFit(lfitdata)
}
\arguments{
\item{lfitdata}{a list of fitted data}
}
\value{
the p-values of the chi-square test and Kolmogorov-Smirnov test
}
\description{
Computes the p-value of the chi-square test and Kolmogorov-Smirnov test and each statistic value
}
\seealso{
Other DistributionAnalysis: \code{\link{cdfcompggplot2}};
  \code{\link{denscompggplot2}}; \code{\link{fitData}};
  \code{\link{qqcompggplot2}}; \code{\link{summaryFit}}
}

