% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/6_DistributionAnalysis.R
\name{goodnessFit}
\alias{goodnessFit}
\title{Computes the p-value of the chi-square test and Kolmogorov-Smirnov test}
\usage{
goodnessFit(lfitdata)
}
\arguments{
\item{lfitdata}{a list of fitted data}
}
\value{
the p-values and the values of the statistics in the chi-square test and Kolmogorov-Smirnov test
}
\description{
Computes the p-value of the chi-square test and Kolmogorov-Smirnov test
}
\examples{
mydata <- rnorm(100, 10, 0.5)

goodnessFit(fitData(mydata))
}
\seealso{
Other DistributionAnalysis: \code{\link{cdfcompggplot2}};
  \code{\link{denscompggplot2}}; \code{\link{fitData}};
  \code{\link{qqcompggplot2}}; \code{\link{summaryFit}}
}

