% Generated by roxygen2 (4.0.2): do not edit by hand
\name{maxCustomers.M_M_1_INF_H}
\alias{maxCustomers}
\alias{maxCustomers.M_M_1_INF_H}
\alias{maxCustomers.M_M_1_K}
\alias{maxCustomers.M_M_S_INF_H}
\alias{maxCustomers.M_M_S_INF_H_Y}
\alias{maxCustomers.M_M_S_K}
\alias{maxCustomers.MarkovianModel}
\title{Returns the maximun value for n that satisfies \ifelse{latex}{P_{n}}{\out{P<sub>n</sub>}} > 0}
\usage{
\method{maxCustomers}{M_M_1_INF_H}(qm)

\method{maxCustomers}{M_M_S_INF_H}(qm)

\method{maxCustomers}{M_M_S_INF_H_Y}(qm)

maxCustomers(qm)

\method{maxCustomers}{MarkovianModel}(qm)

\method{maxCustomers}{M_M_1_K}(qm)

\method{maxCustomers}{M_M_S_K}(qm)
}
\arguments{
\item{qm}{object MarkovianModel}
}
\description{
Returns the maximun value for n that satisfies \ifelse{latex}{P_{n}}{\out{P<sub>n</sub>}} > 0
}
\details{
\code{maxCustomers.M_M_1_INF_H} implements the method for a M/M/1/\eqn{\infty}/H queueing model

\code{maxCustomers.M_M_S_INF_H} implements the method for a M/M/s/\eqn{\infty}/H queueing model

\code{maxCustomers.M_M_S_INF_H_Y} implements the method for a M/M/s/\eqn{\infty}/H/Y queueing model

\code{maxCustomers.MarkovianModel} implements the default method. Returns infinite.

\code{maxCustomers.M_M_1_K} implements the method for a M/M/1/K queueing model

\code{maxCustomers.M_M_S_K} implements the method for a M/M/S/K queueing model
}
\examples{
maxCustomers(M_M_1_INF_H(20, 60, 10))
maxCustomers(M_M_S_INF_H(10, 40, 5, 8))
maxCustomers(M_M_S_INF_H_Y(3, 6, 2, 5, 8))
maxCustomers(M_M_1_K(60, 10, 3))
maxCustomers(M_M_S_K(60, 10, 2, 5))
}

