% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ModelsHeader.R
\name{no_distr}
\alias{no_distr}
\title{Defines an empty object to indicate that there does not exist distribution}
\usage{
no_distr()
}
\description{
Defines an empty object to indicate that there does not exist distribution
}

