\name{node}
\alias{node}
\title{Returns the queueing model which corresponds to the node i of the network}
\usage{
  node(net, i)
}
\arguments{
  \item{net}{Network}

  \item{i}{Node}
}
\value{
  \code{MarkovianModel} object
}
\description{
  Returns the queueing model which corresponds to the node
  i of the network
}

