% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1_ModelsHeader.R
\name{node}
\alias{node}
\title{Returns the queueing model which corresponds to the node i of the Open Jackson Network}
\usage{
node(net, i)
}
\arguments{
\item{net}{Open Jackson Network}

\item{i}{Node}
}
\value{
\code{MarkovianModel} object
}
\description{
Returns the queueing model which corresponds to the node i of the Open Jackson Network
}
\examples{
node(OpenJacksonNetwork(), 1)
}
\keyword{internal}

