\name{P0i}
\alias{P0i}
\alias{P0i.OpenJackson}
\title{Steady-state probability of 0 customers in the system on the node i of an Open Jackson Network.}
\usage{
  P0i(net, i)

  \method{P0i}{OpenJackson} (net, i)
}
\arguments{
  \item{net}{Network}

  \item{i}{Node}
}
\value{
  \deqn{P_{0,i}()}
}
\description{
  Returns the value of the probability of 0 customers in
  node i of an Open Jackson Network.
}
\details{
  \code{P0i.OpenJackson} implements the method for an Open
  Jackson Network model
}

