% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Pi}
\alias{Pi}
\alias{Pi.ClosedJackson}
\alias{Pi.SimulatedNetwork}
\title{Steady-state probability of n customers in the node i of a network.}
\usage{
Pi(net, n, node)

\method{Pi}{ClosedJackson}(net, n, node)

\method{Pi}{SimulatedNetwork}(net, n, node)
}
\arguments{
\item{net}{Network}

\item{n}{Customers}

\item{node}{Node}
}
\value{
\ifelse{latex}{P_{n}}{\out{<i>P<sub>n</sub></i>}} in the selected node
}
\description{
Returns the value \ifelse{latex}{\eqn{P_{i}(n)}}{\out{<i>P<sub>i</sub>(n)</i>}} in the node i of a network
}
\details{
\code{Pi.ClosedJackson} implements the method for a Closed Jackson Network model

\code{Pi.SimulatedNetwork} implements the method for a SimulatedNetwork model
}

