\name{Pi0}
\alias{Pi0}
\alias{Pi0.OpenJackson}
\title{Steady-state probability of i customers in the system on the node 0 of an Open Jackson Network.}
\usage{
  Pi0(net, i)

  \method{Pi0}{OpenJackson} (net, i)
}
\arguments{
  \item{net}{Network}

  \item{i}{Customers}
}
\value{
  \eqn{P_{i,0}()}
}
\description{
  Returns the value of the probability of i customers in
  node 0 of an Open Jackson Network.
}
\details{
  \code{Pi0.OpenJackson} implements the method for an Open
  Jackson Network model
}
\keyword{internal}

