\name{pnlast}
\alias{pnlast}
\title{Probability of i clients in the last node of the network}
\usage{
  pnlast(ps, i)
}
\arguments{
  \item{ps}{partial solution of closed network}

  \item{i}{clients}
}
\value{
  P(i) in the last node
}
\description{
  Probability of i clients in the last node of the network
}
\keyword{internal}

