% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NetworkModels.R
\name{print.ClosedJackson}
\alias{print.ClosedJackson}
\title{Print the main characteristics of a Closed Jackson Network model}
\usage{
\method{print}{ClosedJackson}(x, ...)
}
\arguments{
\item{x}{a ClosedJackson object}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print the main characteristics of a Closed Jackson Network model
}
\keyword{internal}

