\name{print.M_M_INF}
\alias{print.M_M_INF}
\title{Print the main characteristics of a M/M/\eqn{\infty} model}
\usage{
  \method{print}{M_M_INF} (x, ...)
}
\arguments{
  \item{x}{a M_M_INF object}

  \item{...}{Further arguments passed to or from other
  methods.}
}
\description{
  Print the main characteristics of a M/M/\eqn{\infty}
  model
}
\keyword{internal}

