% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1_ModelsHeader.R
\name{print.MarkovianModel}
\alias{print.MarkovianModel}
\title{Print the main characteristics of a queueing model}
\usage{
\method{print}{MarkovianModel}(x, ...)
}
\arguments{
\item{x}{MarkovianModel object}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print the main characteristics of a queueing model
}
\keyword{internal}

