% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Qn.M_M_1_INF_H}
\alias{Qn}
\alias{Qn.M_M_1_INF_H}
\alias{Qn.M_M_1_K}
\alias{Qn.M_M_S_INF_H}
\alias{Qn.M_M_S_INF_H_Y}
\alias{Qn.M_M_S_K}
\alias{Qn.MarkovianModel}
\title{Steady-state probability of finding n customers in the system when a new customer arrives}
\usage{
\method{Qn}{M_M_1_INF_H}(qm, n)

\method{Qn}{M_M_S_INF_H}(qm, n)

\method{Qn}{M_M_S_INF_H_Y}(qm, n)

Qn(qm, n)

\method{Qn}{MarkovianModel}(qm, n)

\method{Qn}{M_M_1_K}(qm, n)

\method{Qn}{M_M_S_K}(qm, n)
}
\arguments{
\item{qm}{Queueing model}

\item{n}{Customers}
}
\value{
\ifelse{latex}{\deqn{Q_n}}{\out{<i>Q<sub>n</sub></i>}}
}
\description{
Returns the probability of n customers in the system in the moment of the
arrival of a customer.
}
\details{
\code{Qn.M_M_1_INF_H} implements the method for a M/M/1/\eqn{\infty}/H queueing model

\code{Qn.M_M_S_INF_H} implements the method for a M/M/s/\eqn{\infty}/H queueing model

\code{Qn.M_M_S_INF_H_Y} implements the method for a M/M/s/\eqn{\infty}/H with Y replacements queueing model

\code{Qn.MarkovianModel} implements the default method (generates a message).

\code{Qn.M_M_1_K} implements the method for a M/M/1/K queueing model

\code{Qn.M_M_S_K} implements the method for a M/M/S/K queueing model
}
\examples{
qm <- M_M_1_INF_H(20, 60, 10)
#Probability of 0,1,2,3,4 and 5 customers in the queue
Qn(qm, 0:5)
qm <- M_M_S_INF_H(10, 40, 5, 8)
#Probability of 0,1,2,3,4 and 5 customers in the queue
Qn(qm, 0:5)
qm <- M_M_S_INF_H_Y(3, 6, 2, 5, 8)
#Probability of 0,1,2,3,4 and 5 customers in the queue
Qn(qm, 0:5)
qm <- M_M_1_K(60, 10, 3)
#Probability of 0,1,2,3,4 and 5 customers in the queue
Qn(qm, 0:5)
qm <- M_M_S_K(60, 10, 2, 5)
#Probability of 0,1,2,3,4 and 5 customers in the system
Qn(qm, 0:5)
}

