% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/6_DistributionAnalysis.R
\name{qqcompggplot2}
\alias{qqcompggplot2}
\title{Q-Q Plot using the package ggplot2}
\usage{
qqcompggplot2(lfitdata)
}
\arguments{
\item{lfitdata}{a list of fitted data}
}
\description{
Q-Q Plot using the package ggplot2
}
\seealso{
Other DistributionAnalysis: \code{\link{cdfcompggplot2}};
  \code{\link{denscompggplot2}}; \code{\link{fitData}};
  \code{\link{goodnessFit}}; \code{\link{summaryFit}}
}

