% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1_ModelsHeader.R
\name{summary.MarkovianModel}
\alias{summary.MarkovianModel}
\title{Shows the main graphics of the parameters of a Markovian Model}
\usage{
\method{summary}{MarkovianModel}(object, t = list(range = seq(object$out$w,
  object$out$w * 3, length.out = 100)), n = c(0:5), only = NULL,
  graphics = "ggplot2", ...)
}
\arguments{
\item{object}{Markovian Model}

\item{t}{Range of t}

\item{n}{Range of n}

\item{only}{String that allows to show only one of the variables. Can be "t" or "n".}

\item{graphics}{Type of graphics: "graphics" uses the basic R plot and "ggplot2" the library ggplot2}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Shows the main graphics of the parameters of a Markovian Model
}

