\name{summary.SimulatedModel}
\alias{summary.SimulatedModel}
\title{Shows the main graphics of the parameters of a Simulated Model}
\usage{
  \method{summary}{SimulatedModel} (object, range = NULL,
    ...)
}
\arguments{
  \item{object}{Simulated Model}

  \item{range}{Range of the graphics}

  \item{...}{Further arguments passed to or from other
  methods.}
}
\description{
  Shows the main graphics of the parameters of a Simulated
  Model
}

