% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/6_DistributionAnalysis.R
\name{summaryFit}
\alias{summaryFit}
\title{Shows three plots:The histogram and theoretical densities, the empirical and theoretical CDF's and the Q-Q plot}
\usage{
summaryFit(lfitdata, graphics = "ggplot2")
}
\arguments{
\item{lfitdata}{a list of fitted data}

\item{graphics}{Type of graphics: "graphics" uses the basic R plot and "ggplot2" the library ggplot2}
}
\description{
Shows three plots:The histogram and theoretical densities, the empirical and theoretical CDF's and the Q-Q plot
}
\seealso{
Other DistributionAnalysis: \code{\link{cdfcompggplot2}};
  \code{\link{denscompggplot2}}; \code{\link{fitData}};
  \code{\link{goodnessFit}}; \code{\link{qqcompggplot2}}
}

