% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1_ModelsHeader.R
\name{summaryPnQn}
\alias{summaryPnQn}
\title{Plot of the probabilities of a Markovian Model}
\usage{
summaryPnQn(object, n, graphics = "ggplot2")
}
\arguments{
\item{object}{Markovian Model}

\item{n}{Range of n}

\item{graphics}{Type of graphics: "graphics" uses the basic R plot and "ggplot2" the library ggplot2}
}
\description{
Shows a plot with the P(n) and Q(n) values of a Markovian Model
}
\keyword{internal}

