\name{summaryPnQn}
\alias{summaryPnQn}
\title{Shows a plot of P(n) and Q(n) values of a Markovian Model}
\usage{
  summaryPnQn(object, n, graphics = "ggplot2")
}
\arguments{
  \item{object}{Markovian Model}

  \item{n}{Range of n}

  \item{graphics}{Type of graphics: "graphics" use the
  basic R plot and "ggplot2" the library ggplot2}
}
\description{
  Shows a plot of P(n) and Q(n) values of a Markovian Model
}

