\name{summaryWtWqt}
\alias{summaryWtWqt}
\title{Shows a plot of W(t) and Wq(t) values of a Markovian Model}
\usage{
  summaryWtWqt(object, t, graphics = "ggplot2")
}
\arguments{
  \item{object}{Markovian Model}

  \item{t}{Range of t}

  \item{graphics}{Type of graphics: "graphics" use the
  basic R plot and "ggplot2" the library ggplot2}
}
\description{
  Shows a plot of W(t) and Wq(t) values of a Markovian
  Model
}

